better test file management and directory structure
